Remove erroneous cast
Don't remember why it's there, it's probably a thinko.

Pointed out by Goswin von Brederlow <goswin-v-b@web.de>

+++ nbd-server.c
@@ -1758,7 +1758,7 @@ int mainloop(CLIENT *client) {
 				continue;
 			}
 
-			if (((ssize_t)((off_t)request.from + len) > client->exportsize)) {
+			if (((off_t)request.from + len) > client->exportsize) {
 				DEBUG("[RANGE!]");
 				ERROR(client, reply, EINVAL);
 				continue;
